Crate pgp

rPGP

rPGP is an OpenPGP implementation.

Usage examples are available under the respective modules: Key generation, signing and verifying with external hashing, packet based signing and verifying.

Re-exports

pub use self::composed::key::*;

pub use self::composed::*;

Modules

armor
Armor module

base64_decoder
base64 decoder module

base64_reader
base64 reader module

composed

crypto
Cryptography module

de
Deserialize trait

errors

line_reader
Line reader module

line_writer
Line writer module

normalize_lines
Line ending normalization module

packet
Packet module

ser
Serialize trait module

types

util
Utilities

Macros

bail

ensure

ensure_eq

err_opt

format_err

impl_public_key

impl_secret_key

impl_try_from_into

key_parser
This macro generates the parsers matching to the two different types of keys, public and secret.

unimplemented_err

unsupported_err

Structs

Signature
Signature Packet https://tools.ietf.org/html/rfc4880.html#section-5.2