Enum pgp::composed::message::Message

pub enum Message {
    Literal(LiteralData),
    Compressed(CompressedData),
    Signed {
        message: Option<Box<Message>>,
        one_pass_signature: Option<OnePassSignature>,
        signature: Signature,
    },
    Encrypted {
        esk: Vec<Esk>,
        edata: Vec<Edata>,
    },
}

An OpenPGP message

Variants

Literal(LiteralData)

Compressed(CompressedData)

Signed

Fields of Signed

message: Option<Box<Message>>

nested message

one_pass_signature: Option<OnePassSignature>

for signature packets that contain a one pass message

signature: Signature

Encrypted

Fields of Encrypted

esk: Vec<Esk>edata: Vec<Edata>

Implementations

impl Message

pub fn new_literal(file_name: &str, data: &str) -> Self

pub fn new_literal_bytes(file_name: &str, data: &[u8]) -> Self

pub fn compress(&self, alg: CompressionAlgorithm) -> Result<Self>

Compresses the message.

pub fn decompress(self) -> Result<Self>

Decompresses the data if compressed.

pub fn encrypt_to_keys<R: CryptoRng + Rng>(
&self,
rng: &mut R,
alg: SymmetricKeyAlgorithm,
pkeys: &[&impl PublicKeyTrait]
) -> Result<Self>

Encrypt the message to the list of passed in public keys.

pub fn encrypt_with_password<R, F>(
&self,
rng: &mut R,
s2k: StringToKey,
alg: SymmetricKeyAlgorithm,
msg_pw: F
) -> Result<Self> where
R: Rng + CryptoRng,
F: FnOnce() -> String + Clone,

Encrytp the message using the given password.

pub fn sign<F>(
self,
key: &impl SecretKeyTrait,
key_pw: F,
hash_algorithm: HashAlgorithm
) -> Result<Self> where
F: FnOnce() -> String,

Sign this message using the provided key.

pub fn into_signature(self) -> StandaloneSignature

Convert the message to a standalone signature according to the cleartext framework.

pub fn verify(&self, key: &impl PublicKeyTrait) -> Result<()>

Verify this message. For signed messages this verifies the signature and for compressed messages they are decompressed and checked for signatures to verify.

pub fn get_recipients(&self) -> Vec<&KeyId>

Returns a list of KeyIds that the message is encrypted to. For non encrypted messages this list is empty.

pub fn decrypt<'a, F, G>(
&'a self,
msg_pw: F,
key_pw: G,
keys: &[&SignedSecretKey]
) -> Result<(MessageDecrypter<'a>, Vec<KeyId>)> where
F: FnOnce() -> String + Clone,
G: FnOnce() -> String + Clone,

Decrypt the message using the given key. Returns a message decrypter, and a list of KeyIds that are valid recipients of this message.

pub fn decrypt_with_password<'a, F>(
&'a self,
msg_pw: F
) -> Result<MessageDecrypter<'a>> where
F: FnOnce() -> String + Clone,

Decrypt the message using the given key. Returns a message decrypter, and a list of KeyIds that are valid recipients of this message.

pub fn is_one_pass_signed(&self) -> bool

Check if this message is a signature, that was signed with a one pass signature.

pub fn is_literal(&self) -> bool

pub fn get_literal(&self) -> Option<&LiteralData>

pub fn get_content(&self) -> Result<Option<Vec<u8>>>

Returns the underlying content and None if the message is encrypted.

pub fn to_armored_writer(
&self,
writer: &mut impl Write,
headers: Option<&BTreeMap<String, String>>
) -> Result<()>

pub fn to_armored_bytes(
&self,
headers: Option<&BTreeMap<String, String>>
) -> Result<Vec<u8>>

pub fn to_armored_string(
&self,
headers: Option<&BTreeMap<String, String>>
) -> Result<String>