Trait pgp::composed::Deserializable

pub trait Deserializable: Sized {
    fn from_packets<'a>(
        packets: impl Iterator<Item = Packet> + 'a
    ) -> Box<dyn Iterator<Item = Result<Self>> + 'a>;

    fn from_bytes(bytes: impl Read) -> Result<Self> { ... }

    fn from_string(input: &str) -> Result<(Self, BTreeMap<String, String>)> { ... }

    fn from_string_many<'a>(
        input: &'a str
    ) -> Result<(Box<dyn Iterator<Item = Result<Self>> + 'a>, BTreeMap<String, String>)> { ... }

    fn from_armor_single<R: Read + Seek>(
        input: R
    ) -> Result<(Self, BTreeMap<String, String>)> { ... }

    fn from_armor_many<'a, R: Read + Seek + 'a>(
        input: R
    ) -> Result<(Box<dyn Iterator<Item = Result<Self>> + 'a>, BTreeMap<String, String>)> { ... }

    fn from_bytes_many<'a>(
        bytes: impl Read + 'a
    ) -> Box<dyn Iterator<Item = Result<Self>> + 'a> { ... }
}

Required methods

fn from_packets<'a>(
packets: impl Iterator<Item = Packet> + 'a
) -> Box<dyn Iterator<Item = Result<Self>> + 'a>

Turn a list of packets into a usable representation.

Loading content...
Provided methods

fn from_bytes(bytes: impl Read) -> Result<Self>

Parse a single byte encoded composition.

fn from_string(input: &str) -> Result<(Self, BTreeMap<String, String>)>

Parse a single armor encoded composition.

fn from_string_many<'a>(
input: &'a str
) -> Result<(Box<dyn Iterator<Item = Result<Self>> + 'a>, BTreeMap<String, String>)>

Parse an armor encoded list of compositions.

fn from_armor_single<R: Read + Seek>(
input: R
) -> Result<(Self, BTreeMap<String, String>)>

Armored ascii data.

fn from_armor_many<'a, R: Read + Seek + 'a>(
input: R
) -> Result<(Box<dyn Iterator<Item = Result<Self>> + 'a>, BTreeMap<String, String>)>

Armored ascii data.

fn from_bytes_many<'a>(
bytes: impl Read + 'a
) -> Box<dyn Iterator<Item = Result<Self>> + 'a>

Parse a list of compositions in raw byte format.

Loading content...