Trait pgp::crypto::hash::Hasher

pub trait Hasher {
    fn update(&mut self, _: &[u8]);

    fn finish(self: Box<Self>) -> Vec<u8>;
}

Trait to work around the fact that the Digest trait from rustcrypto can not be used as Box<Digest>.

Required methods

fn update(&mut self, _: &[u8])

Update the hash with the given value.

fn finish(self: Box<Self>) -> Vec<u8>

Finalize the hash and return the result.

Loading content...